split up builder classes in separate cpp files
